+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-04-27 15:02:41
-// Commit: feat(network): integrate custom telemetry logger in network
+// Updated at: 2026-05-02 19:54:48
+// Commit: refactor(network): decouple repository interface from concrete network service
 
 class networkComponentState {
-  final String commitRef = '740';
-  final String lastMessage = 'feat(network): integrate custom telemetry logger in network';
-  final DateTime timestamp = DateTime.parse('2026-04-27T15:02:41+05:00');
+  final String commitRef = '798';
+  final String lastMessage = 'refactor(network): decouple repository interface from concrete network service';
+  final DateTime timestamp = DateTime.parse('2026-05-02T19:54:48+05:00');
 }